+++ packages/connect-multichain/src/domain/multichain/index.ts
@@ -7,13 +7,13 @@ import type {
 import type { CaipAccountId, Json } from '@metamask/utils';
 
 import { EventEmitter, type SDKEvents } from '../events';
-import type { StoreClient } from '../store/client';
 import type { InvokeMethodOptions, RPCAPI, Scope } from './api/types';
 import type {
   ExtendedTransport,
   MergeableMultichainOptions,
   MultichainOptions,
 } from './types';
+import type { StoreClient } from '../store/client';
 
 export type ConnectionStatus =
   | 'pending'
@@ -44,6 +44,8 @@ export abstract class MultichainCore extends EventEmitter<SDKEvents> {
   abstract transport: ExtendedTransport;
 
   abstract transportType: TransportType;
+
+  abstract version: string;
 
   /**
    * Establishes a connection to the multichain provider, or re-use existing session