+++ packages/connect-evm/CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+<<<<<<< HEAD
+### Added
+
+- Validate `@metamask/connect-multichain` peer version at runtime and warn on mismatch ([#253](https://github.com/MetaMask/connect-monorepo/pull/253))
+
+### Changed
+
+- Move `@metamask/connect-multichain` from `dependencies` to `peerDependencies` ([#253](https://github.com/MetaMask/connect-monorepo/pull/253))
+
+||||||| fcfafa1
+=======
 ## [0.11.1]
 
 ### Fixed
@@ -23,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ensure EIP-1193 provider properties (`selectedChainId`, `accounts`) are updated before emitting the `connect` event ([#269](https://github.com/MetaMask/connect-monorepo/pull/269))
 
+>>>>>>> main
 ## [0.10.0]
 
 ### Changed